Refactor server to connect to Cassandra
