+++ client/src/components/App.jsx
@@ -55,7 +55,7 @@ class App extends React.Component {
   }
 
   getRestaurantPhotos(restaurantId) {
-    axios.get(`http://ec2-34-201-243-233.compute-1.amazonaws.com/restaurant/photos/${restaurantId}`)
+    axios.get(`/api/restaurant/${restaurantId}/photos`)
       .then(results => this.setState({ photos: results.data }))
       .catch(err => console.log('ERROR', err));
   }
