+++ package.json
@@ -35,6 +35,7 @@
     "babel-preset-env": "^1.7.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
+    "cassandra-driver": "^3.5.0",
     "chai": "^4.1.2",
     "css-loader": "^1.0.0",
     "enzyme": "^3.3.0",
