+++ server/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const path = require('path');
-const db = require('../database/database.js');
+const db = require('../database/index.js');
 
 const app = express();
 
@@ -18,7 +18,7 @@ app.use('/restaurant/:id', express.static('public'));
 app.get('/api/restaurant/:id/photos', (req, res) => {
   const id = Number(req.params.id);
   if (id > 0 && id < 100) {
-    db.getPhotos(id, (err, restaurant) => {
+    db.getById(id, (err, restaurant) => {
       if (err) {
         if (err === 404) {
           res.status(404).send('Not Found');
@@ -32,51 +32,51 @@ app.get('/api/restaurant/:id/photos', (req, res) => {
   }
 });
 
-app.post('/api/restaurant/photos', (req, res) => {
-  const newRestaurant = {
-    username: 'Edward Baeg',
-    description: 'This is so yummy',
-    createdAt: new Date(),
-    photoUrls: ['www.google.com'],
-  };
+// app.post('/api/restaurant/photos', (req, res) => {
+//   const newRestaurant = {
+//     username: 'Edward Baeg',
+//     description: 'This is so yummy',
+//     createdAt: new Date(),
+//     photoUrls: ['www.google.com'],
+//   };
 
-  db.addPhotos(newRestaurant)
-    .then(data => {
-      res.status(201).send('processed post request');
-    })
-    .catch(err => {
-      res.status(500).send('error', err);
-    });
+//   db.addPhotos(newRestaurant)
+//     .then(data => {
+//       res.status(201).send('processed post request');
+//     })
+//     .catch(err => {
+//       res.status(500).send('error', err);
+//     });
 
-});
+// });
 
-app.put('/api/restaurant/:id/photos', (req, res) => {
-  const { id } = req.params;
-  const update = {
-    username: 'Brian So',
-    description: 'This was ok',
-    createdAt: new Date(),
-    photoUrls: ['www.wikipedia.org'],
-  };
+// app.put('/api/restaurant/:id/photos', (req, res) => {
+//   const { id } = req.params;
+//   const update = {
+//     username: 'Brian So',
+//     description: 'This was ok',
+//     createdAt: new Date(),
+//     photoUrls: ['www.wikipedia.org'],
+//   };
 
-  db.updatePhotos(id, update, (err) => {
-    if (err) {
-      res.status(500).send('error', err);
-    } else {
-      res.status(201).send(`processed put request for id number ${id}`);
-    }
-  });
-});
+//   db.updatePhotos(id, update, (err) => {
+//     if (err) {
+//       res.status(500).send('error', err);
+//     } else {
+//       res.status(201).send(`processed put request for id number ${id}`);
+//     }
+//   });
+// });
 
-app.delete('/api/restaurant/:id/photos', (req, res) => {
-  const { id } = req.params;
-  db.deletePhotos(id, (err) => {
-    if (err) {
-      res.status(500).send('db error', error);
-    } else {
-      res.status(200).send(`processed delete request for id number ${id}`);
-    }
-  });
-});
+// app.delete('/api/restaurant/:id/photos', (req, res) => {
+//   const { id } = req.params;
+//   db.deletePhotos(id, (err) => {
+//     if (err) {
+//       res.status(500).send('db error', error);
+//     } else {
+//       res.status(200).send(`processed delete request for id number ${id}`);
+//     }
+//   });
+// });
 
 module.exports = app;
